refactor(testing): split multisig and ACL fee tests into per-flavor files
Per review: the multisig fee payment tests move from
fee_payment/singlesig.rs into fee_payment/multisig.rs, and the ACL fee
payment tests (the signature-path test from singlesig.rs and the
exempt-branch test from tests/auth/singlesig_acl.rs) into
fee_payment/singlesig_acl.rs, so each file covers exactly one auth
flavor. The cycle-estimate mirror constants and the shared
assert_single_fee_note helper move to the fee_payment module root; the
ACL component test helpers (mock_component_proc_roots,
compile_call_set_item_script) become pub(super) for reuse.
