+++ crates/miden-testing/tests/auth/singlesig_acl.rs
@@ -17,19 +17,14 @@ use miden_protocol::account::{
     StorageMapKey,
     StorageSlot,
 };
-use miden_protocol::asset::{Asset, AssetAmount, FungibleAsset, TokenSymbol};
+use miden_protocol::asset::{AssetAmount, FungibleAsset, TokenSymbol};
 use miden_protocol::errors::MasmError;
-use miden_protocol::note::{Note, NoteType};
-use miden_protocol::testing::account_id::ACCOUNT_ID_FEE_FAUCET;
+use miden_protocol::note::Note;
 use miden_protocol::testing::storage::MOCK_VALUE_SLOT0;
 use miden_protocol::transaction::{RawOutputNote, TransactionScript};
 use miden_protocol::{Felt, Word};
 use miden_standards::account::access::{Authority, Pausable, PausableManager};
-use miden_standards::account::auth::{
-    AuthSingleSigAcl,
-    FeeConversionInfo,
-    commit_fee_conversion_info,
-};
+use miden_standards::account::auth::AuthSingleSigAcl;
 use miden_standards::account::faucets::{Description, FungibleFaucet, TokenName};
 use miden_standards::account::policies::{
     BurnPolicy,
@@ -38,7 +33,7 @@ use miden_standards::account::policies::{
     TransferPolicy,
 };
 use miden_standards::code_builder::CodeBuilder;
-use miden_standards::note::{BurnNote, TxFeeNote};
+use miden_standards::note::BurnNote;
 use miden_standards::testing::account_component::MockAccountComponent;
 use miden_standards::testing::faucet::user_faucet_single_sig_acl;
 use miden_standards::testing::note::NoteBuilder;
@@ -53,98 +48,6 @@ use crate::prove_and_verify_transaction;
 
 // TESTS
 // ================================================================================================
-
-/// The ACL exempt (no-signature) branch pays the transaction fee in the native fee asset,
-/// ignoring any conversion info committed via the auth args.
-///
-/// With no signature over the transaction summary there is no signer to authorize a
-/// caller-supplied conversion rate, so the exempt branch pays plainly in the kernel-attested
-/// native fee asset at rate 1/1 (mirroring the network account component); paying (rather than
-/// skipping the fee) also means exempt transactions cannot evade fees. This test runs a
-/// state-changing exempt procedure (`set_item`) on a fee-charging chain with no registered
-/// authenticator, supplies an inflated conversion-rate commitment via the auth args, and asserts
-/// a native fee note covering the computed fee is created at rate 1/1 (the inflated rate is
-/// ignored).
-#[tokio::test]
-async fn acl_exempt_branch_pays_native_fee_note() -> anyhow::Result<()> {
-    const VERIFICATION_BASE_FEE: u32 = 500;
-
-    let (_get_item, set_item, _account_procedure_1) = mock_component_proc_roots();
-
-    let fee_faucet_id = ACCOUNT_ID_FEE_FAUCET.try_into()?;
-    let fee_asset = FungibleAsset::new(fee_faucet_id, 1_000_000)?;
-
-    // `set_item` is exempt, so calling it takes the no-signature branch while still changing state
-    // (making the transaction valid without requiring a signature).
-    let component: AccountComponent =
-        MockAccountComponent::with_slots(AccountStorage::mock_storage_slots()).into();
-    let (auth_component, _authenticator) = Auth::Acl {
-        exempt_procedures: BTreeSet::from([set_item]),
-        auth_scheme: AuthScheme::Falcon512Poseidon2,
-    }
-    .build_component();
-
-    let account = AccountBuilder::new([0; 32])
-        .with_auth_component(auth_component)
-        .with_component(component)
-        .account_type(AccountType::Public)
-        .with_assets([fee_asset.into()])
-        .build_existing()?;
-
-    let mut builder = MockChain::builder().verification_base_fee(VERIFICATION_BASE_FEE);
-    builder.add_account(account.clone())?;
-    let mock_chain = builder.build()?;
-
-    // an inflated conversion rate that would double the paid amount if the exempt branch honored
-    // the committed conversion info
-    let conversion_info = FeeConversionInfo::new(fee_faucet_id, 2, 1)?;
-    let (args, advice_value) =
-        commit_fee_conversion_info(conversion_info, Word::from([9u32, 10, 11, 12]));
-
-    // no authenticator is registered, so a successful execution proves the exempt (no-signature)
-    // branch ran
-    let executed = mock_chain
-        .build_tx_context(account.id(), &[], &[])?
-        .tx_script(compile_call_set_item_script()?)
-        .auth_args(args)
-        .extend_advice_map([(args, advice_value)])
-        .build()?
-        .execute()
-        .await?;
-
-    // exactly one output note is created: a public TX_FEE note carrying the native fee asset
-    assert_eq!(executed.output_notes().num_notes(), 1);
-    let output_note = executed.output_notes().get_note(0);
-    assert_eq!(output_note.metadata().tag(), TxFeeNote::TAG);
-    assert_eq!(output_note.metadata().note_type(), NoteType::Public);
-
-    let assets = output_note.assets();
-    assert_eq!(assets.num_assets(), 1);
-    let asset = assets.iter().next().expect("fee note should carry an asset");
-    let Asset::Fungible(paid_asset) = asset else {
-        panic!("fee note asset should be fungible");
-    };
-    assert_eq!(paid_asset.faucet_id(), fee_faucet_id);
-
-    // the paid amount covers the computed fee at rate 1/1 with a bounded overshoot; the inflated
-    // 2/1 rate committed via the auth args was ignored (honoring it would have doubled the
-    // amount, far exceeding the bound)
-    let required_fee = executed.compute_fee();
-    assert!(
-        paid_asset.amount() >= required_fee,
-        "paid fee {} should cover the required fee {required_fee}",
-        paid_asset.amount()
-    );
-    let max_overpayment = u64::from(3 * VERIFICATION_BASE_FEE);
-    assert!(
-        paid_asset.amount().as_u64() <= required_fee.as_u64() + max_overpayment,
-        "paid fee {} should not exceed the required fee {required_fee} by more than \
-         {max_overpayment}",
-        paid_asset.amount()
-    );
-
-    Ok(())
-}
 
 #[rstest]
 #[case::ecdsa(AuthScheme::EcdsaK256Keccak)]
@@ -593,8 +496,8 @@ struct TestSetup {
 
 /// Returns the procedure roots used by the ACL tests, in this order:
 ///   (`get_item`, `set_item`, `account_procedure_1`).
-fn mock_component_proc_roots() -> (AccountProcedureRoot, AccountProcedureRoot, AccountProcedureRoot)
-{
+pub(super) fn mock_component_proc_roots()
+-> (AccountProcedureRoot, AccountProcedureRoot, AccountProcedureRoot) {
     let component: AccountComponent =
         MockAccountComponent::with_slots(AccountStorage::mock_storage_slots()).into();
 
@@ -674,7 +577,7 @@ fn compile_call_get_item_script() -> anyhow::Result<TransactionScript> {
 
 /// Compiles the canonical "call `mock::account::set_item` on `MOCK_VALUE_SLOT0` with a fixed
 /// dummy word" tx script.
-fn compile_call_set_item_script() -> anyhow::Result<TransactionScript> {
+pub(super) fn compile_call_set_item_script() -> anyhow::Result<TransactionScript> {
     let src = format!(
         r#"
         use mock::account
